virtme-ng: allow to run snaps inside virtme-ng instances
Enable all the required (kernel and user-space) features to start snapd
and execute snaps directly inside a virtme-ng instance.

The snap support is activated only if snapd is installed in the host
system. This won't introduce any additional overhead in systems that
don't have snap available.

In systems that have snapd available, virmte-ng will automatically start
it at boot (tricking it a bit into believing that it's running in a
standard systemd-enabled environment) and at this point we can start any
snap as regular system binaries inside virtme-ng.

This fixes issue #21.

Open issues
===========

The main issue at the moment is that the guest requires read access to
/var/lib/snapd/state.json that has 0600 permissions by default. Since we
are running the guest as unprivileged user, by default it won't have
access to the snapd state, therefore snapd cannot be started in the
guest.

For this reason in order to properly support snapd inside virtme-ng
we need to change the permissions mask of this file and make it readable
by everyone (unless we figure out a better solution to this).

virtme-ng will print a warning in this case, to inform the user to
run chmod manually (sad...) to change the permission of state.json.

However, with this change applied we can run any kind of snap inside
virtme-ng, even graphic applications (using `--graphics`), such as
videogames, web browsers, etc.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>