virtme-ng-init: initialize XDG_RUNTIME_DIR
XDG_RUNTIME_DIR is an environment variable that specifies the path to a
directory where user-specific runtime files should be stored. It is
defined by the X Desktop Group (XDG) Base Directory Specification, which
is a set of guidelines for storing user-specific files in a standardized
manner.

Initialize this working directory and the corresponding environment
variable when the graphic environment is enabled.

This allows better compatibility with certain graphic applications (such
as web browsers, videogames, etc.).

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>