Simplify an env var check
Instead of checking if the env variable is set and then early-returning,
this commit reverses the order so that the function is only run if the
variable is not present.

Also, instead of pattern matching with "if let", this uses is_err() to
check if the variable is not set.

Signed-off-by: Uli Schlachter <psychon@znc.in>