Simplify(?) some environment lookup
The code previously initiated the "uid" variable to a default value and
then had two nested "if"s that both must succeed for another value to be
set.

In this commit, I change that to a processing chain on Option: If
env::var() produces a value, Option::and_then() is used to call another
function. If either of these two steps produces a None, the result will
be a None. .unwrap_or() is then used to replace the None with a default
value.

This allows to get rid of the "mut" on the variable.

Signed-off-by: Uli Schlachter <psychon@znc.in>