Use uname() from nix instead of libc
This allows to get rid of all the unsafety and the conversion to CStr
(well, OsStr now) is handled by nix for us.

I also delayed the call to into_owned() a bit. This is used to convert a
Cow<'_, str> to String, but the "show_machine" case does not actually
need it. Thus, this removes a memory allocation.

Signed-off-by: Uli Schlachter <psychon@znc.in>