Allow run_cmd() with non-utf8-strings
The argument to Command::new() in the Rust standard library is anything
that implements AsRef<OsStr>. However, the wrapper function run_cmd() so
far only allows &str. This meant that run_udevd() had to use
to_string_lossy() to turn a PathBuf into a String that could be passed
to this function.

This commit changes run_cmd() to accept an AsRef<OsStr> so that the call
to to_string_lossy() is no longer necessary. The implementation then
could no longer use format!() with {} to display the value since OsStr
does not implement Display. Thus, I switched this to use the Debug
formatting, but I did not actually check how that looks like.

This change then triggered a clippy warning that I am fixing in this
commit as well:

warning: the borrowed expression implements the required traits
   --> src/main.rs:813:32
    |
813 |                 utils::run_cmd(&format!("{}/virtme-snapd-script", guest_tools_dir), &[]);
    |                                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: change this to: `format!("{}/virtme-snapd-script", guest_tools_dir)`
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#needless_borrow
    = note: `#[warn(clippy::needless_borrow)]` on by default

Signed-off-by: Uli Schlachter <psychon@znc.in>