Avoid allocations for static strings for run_shell()
So far, the code calling run_shell() constructed a Vec<String> just
because one of the entries needed to be a dynamic string. This commit
changes that to be a Vec<&str>. The one dynamic argument is allocated to
another variable and then just borrowed. This allows to get rid of some
to_owned() calls, which IMO make the code more readable.

Then, run_shell() does not actually need a Vec as its argument, so this
is changed into a slice that can then also be passed directly to
Commands::args().

Signed-off-by: Uli Schlachter <psychon@znc.in>