Add guest hostname to /etc/hosts
Some programs (e.g. 'sudo' in some configurations) apparently get
confused if they can't resolve the hostname.  To avoid such problems,
if a hostname for the guest has been specified, bind-mount a
replacement version of /etc/hosts with a pair of entries (IPv4 and
IPv6) added for the given hostname.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>