virtme-ng-init: channel the return code of a command to the host
Send the return code of the command executed inside the guest to the
host.

This allows to run commands inside a guest and give users the impression
that they are running on the host.

This feature is particularly useful for automation with virtme-ng, since
it easily allows to check the return code of a script executed inside
virtme-ng as if it was running directly on the host.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>