net: support multiple interfaces
virtme-run supports multiple --net arguments to setup virtio-net
devices, but the init program was only starting the DHCP client for one
of them: the first one it discovered, which was not necessarily the
first one by alphabetical order.

Now, the DHCP client is started for each interface, so each of them can
have an assigned IP address.

Technically, instead of returning one network device, a vector of
devices is returned. Same for the threads to start the DHCP clients.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>